+++ plugin-explorer/src/app/utilities/messages.ts
@@ -1,0 +1,11 @@
+import { vscode } from './vscode';
+
+export enum MessageType {
+  getPlugins = 'getPlugins',
+  getPlugin = 'getPlugin',
+  getInstalledDeps = 'getInstalledDeps',
+}
+
+export function sendMessage(command: MessageType, value: string) {
+  vscode.postMessage({ command, text: value });
+}