+++ src/rules-capacitor.ts
@@ -117,6 +117,12 @@ export function checkCapacitorRules(project: Project) {
     'This plugin appears to have been abandoned in 2015. Consider using ion-datetime.'
   );
 
+  project.recommendRemove(
+    '@jcesarmobile/ssl-skip',
+    '@jcesarmobile/ssl-skip',
+    'This plugin should only be used during development. Submitting an app with it included will cause it to be rejected.'
+  );
+
   if (exists('cordova-plugin-file-transfer') && !exists('cordova-plugin-whitelist')) {
     // Latest 1.7.1 of the file-transfer plugin requires whitelist in Capacitor projects. See: https://github.com/ionic-team/capacitor/issues/1199
     project.recommendAdd(
@@ -352,7 +358,6 @@ export function capacitorRecommendations(project: Project, forMigration: boolean
   tips.push(notRequiredPlugin('cordova-support-google-services'));
   tips.push(incompatiblePlugin('cordova-plugin-passbook'));
   tips.push(incompatibleReplacementPlugin('cordova-plugin-ionic-keyboard', '@capacitor/keyboard'));
-  tips.push(incompatibleReplacementPlugin('cordova-plugin-splashscreen', '@capacitor/splash-screen'));
 
   tips.push(
     incompatiblePlugin(