+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### Version 1.36
+
+- NX projects run web will use external address and work with Nexus Browser
+- On Windows machines starters will correctly suggest a project name
+- Running iOS and Android projects at the same time no longer queues
+- Fix for NX projects and Capacitor detection
+- Add jetifier removal recommendation
+
 ### Version 1.35
 
 - Fix live reload for Angular project with non default (app) project names